add schedue job to detect expire date of each ticket to delete it

+++ app/Http/Controllers/CategoriesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Category;
+use Auth;
 
 class CategoriesController extends Controller
 {
@@ -19,32 +20,55 @@ class CategoriesController extends Controller
     }
 
   public function create(){
+    if(Auth::user()->hasrole('admin')){
     return view('admin.categories.create');
 
+    }
+
+    return view('not-authorize');
 }
 public function store(Request $request){
-    Category::create([
-        'name' => $request->name
-    ]);
-    return redirect('/admin/categories');
-}
-public function edit($id){
-    $category = Category::find($id);
-    return view('admin.categories.edit' , compact('category'));
+    if(Auth::user()->hasrole('admin')){
+        Category::create([
+            'name' => $request->name
+        ]);
+        return redirect('/admin/categories');
+    }
+    return view('not-authorize');
 
 }
+public function edit($id){
+    if(Auth::user()->hasrole('admin')){
+        $category = Category::find($id);
+        return view('admin.categories.edit' , compact('category'));
+
+    }
+    return view('not-authorize');
+}
   public function show($id){
-    $category = Category::find($id);
-    return view('admin.categories.show' , compact('category'));
+    if(Auth::user()->hasrole('admin')){
+        $category = Category::find($id);
+        return view('admin.categories.show' , compact('category'));
+    }
+    return view('not-authorize');
+
 }
     public function update(Request $request ,$id){
+    if(Auth::user()->hasrole('admin')){
         $getCategory = Category::find($id);
         $getCategory->name = $request->name;
         $getCategory->save();
         return redirect('/admin/categories');
     }
+        return view('not-authorize');
+    }
     public function delete($id){
-        Category::find($id)->delete();
-        return redirect('/admin/categories');
+        if(Auth::user()->hasrole('admin')){
+            Category::find($id)->delete();
+            return redirect('/admin/categories');
+        }
+        return view('not-authorize');
+
+
     }
 }
